movement not dependant on FPS now

+++ src/Collision.cpp
@@ -55,7 +55,7 @@ bool checkCollision(Circle& a, Circle& b) {
 */
 bool checkCollision(Circle& a, SDL_Rect& b) {
 	//Closest point on collision box
-	int cX, cY;
+	float cX, cY;
 
 	//Find closest x offset
 	if (a.x < b.x) cX = b.x;
@@ -71,8 +71,8 @@ bool checkCollision(Circle& a, SDL_Rect& b) {
 	return false;
 }
 
-double distanceSquared(int x1, int y1, int x2, int y2) {
-	int deltaX = x2 - x1;
-	int deltaY = y2 - y1;
+float distanceSquared(float x1, float y1, float x2, float y2) {
+	float deltaX = x2 - x1;
+	float deltaY = y2 - y1;
 	return deltaX * deltaX + deltaY * deltaY;
 }
