+++ src/Collision.h
@@ -8,5 +8,5 @@
 bool checkCollision(SDL_Rect& a, SDL_Rect& b);
 bool checkCollision(Circle& a, Circle& b);
 bool checkCollision(Circle& a, SDL_Rect& b);
-double distanceSquared(int x1, int y1, int x2, int y2);
+float distanceSquared(float x1, float y1, float x2, float y2);
 #endif
